Move the AutoloaderBootstrap to its own class :snake:
